+++ docs/source/changelog.rst
@@ -13,6 +13,20 @@ CT
   module. This has been fixed.
 
 
+VMAT
+^^^^
+`
+* The standard deviation for each VMAT segment is now available as the ``.stdev`` property.
+  The value is expressed as a percentage to be consistent with :math:`R_{corr}`. I.e.
+  a value of 0.5 means 0.5% standard deviation of the ratio of DMLC and open field over the ROI.
+
+  .. code-block:: python
+
+    vmat = DRMLC(...)
+    vmat.analyze(...)
+    data = vmat.results_data()
+    print(data.segments[0].stdev)  # first segment stdev
+
 v 3.14.0
 --------